chore. 오늘의 칭찬 뷰에 필사 뷰 네비게이션 연결
